修复网易云歌曲ID溢出问题

+++ src/ZonyLrcTools.Common/Lyrics/Providers/NetEase/JsonModel/GetSongDetailsRequest.cs
@@ -4,13 +4,13 @@ namespace ZonyLrcTools.Common.Lyrics.Providers.NetEase.JsonModel
 {
     public class GetSongDetailsRequest
     {
-        public GetSongDetailsRequest(int songId)
+        public GetSongDetailsRequest(long songId)
         {
             SongId = songId;
             SongIds = $"%5B{songId}%5D";
         }
 
-        [JsonProperty("id")] public int SongId { get; }
+        [JsonProperty("id")] public long SongId { get; }
 
         [JsonProperty("ids")] public string SongIds { get; }
     }
